requirements/requirements.txt
envisage==4.6.0
click==6.7
six==1.10.0
stevedore==1.24.0
pyzmq==16.0.2
